requirements.txt
pandas>=2.0.0
scikit-learn>=1.0.0
ta>=0.10.0
pytest>=8.0.0
pytest-cov>=6.0.0
coverage>=7.0.0
numpy>=1.22.0
matplotlib>=3.5.0
tensorflow>=2.12.0
keras_tuner>=1.1.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0 
tqdm>=4.67.0
ta
xgboost
git+https://github.com/Vaquum/Binancial
python-dotenv